+++ Tests/xcodeprojTests/Objects/Project/PBXProjectTests.swift
@@ -30,4 +30,117 @@ final class PBXProjectTests: XCTestCase {
         ]
         XCTAssertEqual(attributes, expectedAttributes)
     }
+    
+    func test_addLocalSwiftPackage() throws {
+        // Given
+        let objects = PBXObjects(objects: [])
+
+        let buildPhase = PBXFrameworksBuildPhase(
+            files: [],
+            inputFileListPaths: nil,
+            outputFileListPaths: nil, buildActionMask: PBXBuildPhase.defaultBuildActionMask,
+            runOnlyForDeploymentPostprocessing: true
+        )
+        let target = PBXNativeTarget(name: "Target",
+                                     buildConfigurationList: nil,
+                                     buildPhases: [buildPhase])
+        objects.add(object: target)
+        
+        let configurationList = XCConfigurationList.fixture()
+        let mainGroup = PBXGroup.fixture()
+        objects.add(object: configurationList)
+        objects.add(object: mainGroup)
+
+        let project = PBXProject(name: "Project",
+                                 buildConfigurationList: configurationList,
+                                 compatibilityVersion: "0",
+                                 mainGroup: mainGroup,
+                                 targets: [target])
+
+        objects.add(object: project)
+        
+        // When
+        let packageProduct = try project.addLocalSwiftPackage(path: "Product",
+                                         productName: "Product",
+                                         targetName: target.name)
+        
+        // Then
+        XCTAssertEqual(packageProduct, objects.buildFiles.first?.value.product)
+        XCTAssertEqual(packageProduct, objects.swiftPackageProductDependencies.first?.value)
+        XCTAssertEqual(packageProduct, target.packageProductDependencies.first)
+        
+        XCTAssertEqual(objects.fileReferences.first?.value.name, "Product")
+        
+        XCTAssertEqual(objects.swiftPackageProductDependencies.first?.value, buildPhase.files?.first?.product)
+    }
+    
+    func test_addLocalSwiftPackage_throws_frameworksPhaseError() {
+        // Given
+        let objects = PBXObjects(objects: [])
+
+        let target = PBXNativeTarget(name: "Target")
+        objects.add(object: target)
+        
+        let configurationList = XCConfigurationList.fixture()
+        let mainGroup = PBXGroup.fixture()
+        objects.add(object: configurationList)
+        objects.add(object: mainGroup)
+        
+        let project = PBXProject(name: "Project",
+                                 buildConfigurationList: configurationList,
+                                 compatibilityVersion: "0",
+                                 mainGroup: mainGroup,
+                                 targets: [target])
+        
+        objects.add(object: project)
+        
+        // Then
+        
+        XCTAssertThrowsSpecificError(try project.addLocalSwiftPackage(path: "Product",
+                                                                      productName: "Product",
+                                                                      targetName: target.name),
+                                     PBXProjError.frameworksBuildPhaseNotFound(targetName: target.name))
+    }
+    
+    func test_addSwiftPackage() throws {
+        // Given
+        let objects = PBXObjects(objects: [])
+        
+        let buildPhase = PBXSourcesBuildPhase(
+            files: [],
+            inputFileListPaths: nil,
+            outputFileListPaths: nil, buildActionMask: PBXBuildPhase.defaultBuildActionMask,
+            runOnlyForDeploymentPostprocessing: true
+        )
+        let target = PBXNativeTarget(name: "Target",
+                                     buildConfigurationList: nil,
+                                     buildPhases: [buildPhase])
+        objects.add(object: target)
+        
+        let configurationList = XCConfigurationList.fixture()
+        let mainGroup = PBXGroup.fixture()
+        objects.add(object: configurationList)
+        objects.add(object: mainGroup)
+        
+        let project = PBXProject(name: "Project",
+                                 buildConfigurationList: configurationList,
+                                 compatibilityVersion: "0",
+                                 mainGroup: mainGroup,
+                                 targets: [target])
+        
+        objects.add(object: project)
+        
+        // When
+        let remoteReference = try project.addSwiftPackage(repositoryURL: "url",
+                                                         productName: "Product",
+                                                         versionRequirement: .branch("master"),
+                                                         targetName: "Target")
+        
+        // Then
+        XCTAssertEqual(remoteReference, project.packages.first)
+        XCTAssertEqual(remoteReference, objects.remoteSwiftPackageReferences.first?.value)
+        
+        XCTAssertEqual(remoteReference, objects.buildFiles.first?.value.product?.package)
+        XCTAssertEqual(objects.swiftPackageProductDependencies.first?.value, buildPhase.files?.first?.product)
+    }
 }